// frontEnd/friendFinder/src/app/core/models/InterestCategory.ts
export enum InterestCategory {
  SPORTS = 'SPORTS',
  TECHNOLOGY = 'TECHNOLOGY',
  ARTS = 'ARTS',
  MUSIC = 'MUSIC',
  MOVIES = 'MOVIES',
  BOOKS = 'BOOKS',
  TRAVEL = 'TRAVEL',
  FOOD = 'FOOD',
  GAMING = 'GAMING',
  FITNESS = 'FITNESS',
  PHOTOGRAPHY = 'PHOTOGRAPHY',
  FASHION = 'FASHION',
  EDUCATION = 'EDUCATION',
  BUSINESS = 'BUSINESS',
  SCIENCE = 'SCIENCE',
  OTHER = 'OTHER'
}